requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
anytree==2.12.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cmeel==0.53.3
cmeel-assimp==5.3.1
cmeel-boost==1.83.0
cmeel-console-bridge==1.0.2.2
cmeel-octomap==1.9.8.2
cmeel-qhull==8.0.2.1
cmeel-tinyxml==2.6.2.3
cmeel-urdfdom==3.1.1.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dex_retargeting==0.4.5
docker-pycreds==0.4.0
eigenpy==3.5.1
evdev==1.7.1
executing==2.1.0
Farama-Notifications==0.0.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gym==0.23.1
gym-notices==0.0.8
gymnasium==1.0.0
hpp-fcl==2.4.4
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
# Editable install with no version control (isaacgym==1.0rc4)
-e /home/seongsu/isaacgym/python
isaacgym-stubs==1.1rc4
jax==0.4.13
jaxlib==0.4.13
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.2
nlopt==2.7.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pin==2.7.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pynput==1.7.7
pyparsing==3.1.4
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
python-xlib==0.33
pytransform3d==3.9.0
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sounddevice==0.5.1
stack-data==0.6.3
sympy==1.13.3
tomli==2.2.1
torch==2.4.1
torchvision==0.19.1
traitlets==5.14.3
transform3d==0.0.4
transforms3d==0.4.2
trimesh==4.5.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
